WebApp/requirement.txt
bokeh==3.0.2
brotlipy==0.7.0
contourpy==1.0.6
cycler==0.11.0
dnspython==2.2.1
email-validator==1.3.0
Flask==2.2.2
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.0.1
fonttools==4.29.0
greenlet==2.0.1
kiwisolver==1.3.2
matplotlib==3.5.1
numpy==1.22.1
opencv-python==4.5.5.62
packaging==21.3
pandas==1.4.0
Pillow==9.0.0
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
scipy==1.7.3
seaborn==0.11.2
six==1.16.0
SQLAlchemy==1.4.45
torch==1.12.1
torchaudio==0.12.1
torchvision==0.13.1
Werkzeug==2.2.2
wincertstore==0.2
WTForms==3.0.1
xyzservices==2022.9.0
